requirements.txt
Django==4.2.5
plaid-python==9.0.0
requests==2.31.0
urllib3==2.0.4